Use short data type

+++ edge_tangle/utils/qubo_utils.py
@@ -82,7 +82,7 @@ def qubo_matrix_from_graph(graph: nx.DiGraph, alpha: float | None=None) -> tuple
     lambda_g = 2
     lambda_w = 1
 
-    qubo_matrix = np.zeros((T_max, V + 1, T_max, V + 1))
+    qubo_matrix = np.zeros((T_max, V + 1, T_max, V + 1), dtype=np.int8)
     
     # Path
     for t in range(T_max):
